Move bookmarks' production code to components/bookmarks
Move all the production code of bookmarks to the components folder as
they form a cyclic graph and moving them separately would be more work
for little gain.

BUG=359565, 360637, 360640, 360643, 364881
TBR=willchan

Review URL: https://codereview.chromium.org/253753005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267046 0039d316-1c4b-4281-b951-d872f2087c98

+++ components/bookmarks/core/browser/bookmark_index.cc
@@ -1,9 +1,9 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2014 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "chrome/browser/bookmarks/bookmark_index.h"
+#include "components/bookmarks/core/browser/bookmark_index.h"
 
 #include <algorithm>
 #include <functional>
@@ -14,10 +14,10 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/logging.h"
 #include "base/strings/string16.h"
 #include "base/strings/utf_offset_string_conversions.h"
-#include "chrome/browser/bookmarks/bookmark_utils.h"
 #include "components/bookmarks/core/browser/bookmark_client.h"
 #include "components/bookmarks/core/browser/bookmark_match.h"
 #include "components/bookmarks/core/browser/bookmark_node.h"
+#include "components/bookmarks/core/browser/bookmark_utils.h"
 #include "components/query_parser/query_parser.h"
 #include "components/query_parser/snippet.h"
 #include "third_party/icu/source/common/unicode/normalizer2.h"
@@ -33,7 +33,8 @@ base::string16 Normalize(const base::string16& text) {
   UErrorCode status = U_ZERO_ERROR;
   const icu::Normalizer2* normalizer2 =
       icu::Normalizer2::getInstance(NULL, "nfkc", UNORM2_COMPOSE, status);
-  icu::UnicodeString unicode_text(text.data(), text.length());
+  icu::UnicodeString unicode_text(
+      text.data(), static_cast<int32_t>(text.length()));
   icu::UnicodeString unicode_normalized_text;
   normalizer2->normalize(unicode_text, unicode_normalized_text, status);
   if (U_FAILURE(status))
